requirements.txt
# requirements.txt
Django==4.2.7
djangorestframework==3.14.0
phonenumbers==8.13.25
twilio==8.10.0
requests==2.31.0
python-decouple==3.8